# Implement rendering optimization in the Scheduler Track