@karupayun 2022-09-07
> Lgtm, thanks a lot for this. A suggestion, feel free to ignore or do this in a follow-up PR: Depending on how soon you want to merge the different implementations, would it make sense to leave `TimerTrack::DoUpdatePrimitives` empty and instead provide something like `TimerTrack::RenderTimersUnoptimized`? Then you could call this in all timer track subclasses that do not implement any special optimization and it would be easier to follow which class uses which rendering code instead of omitting the `super` calls.

I will include this comment in the tracking bug. This was kind of an experiment in SchedulerTrack, but we need to provide two methods to render timers, I agree.

I'm not sure if it's not better to have the logic directly in TimerTracks for now. TimerTrack::DoUpdatePrimitives could use internal methods (like UpdateOrderedPrimitives or UpdatePrimitivesUnoptimized) depending on the type of the Track, but maybe you are right and this can be chosen in the subclasses and there it can be better explained why we are "not optimizing". 

Anyway, now we have Scheduler as the exception and I prefer this than having 6 different exceptions. As soon as this is implemented, we can solve this tech-debt. For now, I think that having Scheduler as the exception is the best way.